requirements.txt
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
JSON-log-formatter==0.3.0
mongoengine==0.19.1
pycparser==2.20
pymongo==3.10.1
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.4
